native-image work, bug fixes

+++ src/main/java/com/ebremer/beakgraph/hdf5/readers/FiveSectionDictionaryReader.java
@@ -147,7 +147,7 @@ public class FiveSectionDictionaryReader implements GSPODictionary {
     }
 
     @Override
-    public Object extractGraph(int id) {
+    public Object extractGraph(long id) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
@@ -157,7 +157,7 @@ public class FiveSectionDictionaryReader implements GSPODictionary {
     }
 
     @Override
-    public Object extractSubject(int id) {
+    public Object extractSubject(long id) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
@@ -167,7 +167,7 @@ public class FiveSectionDictionaryReader implements GSPODictionary {
     }
 
     @Override
-    public Object extractPredicate(int id) {
+    public Object extractPredicate(long id) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 
@@ -177,7 +177,7 @@ public class FiveSectionDictionaryReader implements GSPODictionary {
     }
 
     @Override
-    public Object extractObject(int id) {
+    public Object extractObject(long id) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
